maintenance: move files to subdirectory
