+++ ydb/core/grpc_services/ya.make
@@ -5,7 +5,7 @@ SRCS(
     audit_dml_operations.cpp
     db_metadata_cache.h
 
-    etcd/api/etcdserverpb/rpc_range.cpp
+    etcd/api/kv/rpc_range.cpp
 
     grpc_endpoint_publish_actor.cpp
     grpc_helper.cpp
